Updated exif data to use Exiflib for faster reading
Using Exiflib now so I don't have to initialize GDI objects.

+++ AutoRotateImages/Program.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using System.IO;
 using System.Drawing;
 using System.Drawing.Imaging;
+using ExifLib;
 
 namespace AutoRotateImages
 {
@@ -23,55 +24,59 @@ namespace AutoRotateImages
 
             foreach (var item in files)
             {
-                FileStream fs = new FileStream(item, FileMode.Open);
-                Image img = Image.FromStream(fs);
-                PropertyItem pi = img.GetPropertyItem(274);
-                int orientation = pi.Value[0];
-                EncoderParameters encoderParameters = new EncoderParameters(1);
-                
-                bool edited = false;
-                
-
-                switch (orientation)
+                using (ExifReader reader = new ExifReader(item))
                 {
-                    case 1:
-                        continue;
-                    
-                    case 6:
-                        encoderParameters.Param[0] = new EncoderParameter(System.Drawing.Imaging.Encoder.Transformation, (long)EncoderValue.TransformRotate90);
-                        edited = true;
-                        break;
-                    case 8:
-                        encoderParameters.Param[0] = new EncoderParameter(System.Drawing.Imaging.Encoder.Transformation, (long)System.Drawing.Imaging.EncoderValue.TransformRotate270);
-                        edited = true;
-                        break;
-                }
+                    UInt16 orientation;
+                    reader.GetTagValue(ExifTags.Orientation, out orientation);
 
-                if (edited)
-                {
-                    // Set orientation to normal so if we run this again it doesn't rotate an already rotated image
-                    pi.Value[0] = 1;
-                    img.SetPropertyItem(pi);
-
-                    // Added new memorystream
-                    MemoryStream ms = new MemoryStream();
-
-                    // Read into memorystream
-                    img.Save(ms, GetEncoder(img.RawFormat), encoderParameters);
-                    fs.Close();
-                    File.Delete(item);
-
-                    // Read from memorystream into byte array
-                    byte[] ni = ms.ToArray();
-
-                    // Save out using new filestream.
-                    using (FileStream nfs = new FileStream(item, FileMode.Create, FileAccess.ReadWrite))
+                    if(orientation != 1)
                     {
-                        nfs.Write(ni, 0, ni.Length);
+                        EncoderParameters encoderParameters = new EncoderParameters(1);
+                        FileStream fs = new FileStream(item, FileMode.Open);
+                        Image img = Image.FromStream(fs);
+                        PropertyItem pi = img.GetPropertyItem(274);
+
+                        bool edited = false;
+
+                        switch (orientation)
+                        {
+                            // TODO: Handle other orientations and flips
+                            case 6:
+                                encoderParameters.Param[0] = new EncoderParameter(System.Drawing.Imaging.Encoder.Transformation, (long)EncoderValue.TransformRotate90);
+                                edited = true;
+                                break;
+                            case 8:
+                                encoderParameters.Param[0] = new EncoderParameter(System.Drawing.Imaging.Encoder.Transformation, (long)System.Drawing.Imaging.EncoderValue.TransformRotate270);
+                                edited = true;
+                                break;
+                        }
+
+                        if (edited)
+                        {
+                            // Set orientation to normal so if we run this again it doesn't rotate an already rotated image
+                            pi.Value[0] = 1;
+                            img.SetPropertyItem(pi);
+
+                            // Added new memorystream
+                            MemoryStream ms = new MemoryStream();
+
+                            // Read into memorystream
+                            img.Save(ms, GetEncoder(img.RawFormat), encoderParameters);
+                            reader.Dispose();
+                            fs.Close();
+                            File.Delete(item);
+
+                            // Read from memorystream into byte array
+                            byte[] ni = ms.ToArray();
+
+                            // Save out using new filestream.
+                            using (FileStream nfs = new FileStream(item, FileMode.Create, FileAccess.ReadWrite))
+                            {
+                                nfs.Write(ni, 0, ni.Length);
+                            }
+                        }
                     }
                 }
-
-
             }
         }
